% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMixtReg-package.R
\docType{package}
\name{BayesMixtReg-package}
\alias{BayesMixtReg}
\alias{BayesMixtReg-package}
\title{BayesMixtReg: Bayesian mixture of regressions}
\description{
Bayesian mixture of regressions.
}
\keyword{internal}
