% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.logisticRWMHcond}
\alias{.logisticRWMHcond}
\title{Logistic binomial regression / conditional inference with RWMH method}
\usage{
.logisticRWMHcond(
  N,
  nbAdaptMarche,
  y,
  X,
  maxBinom,
  curr,
  SigMarch,
  prior,
  warmup = 0L,
  invTemp = 1
)
}
\arguments{
\item{N}{the number of iterations post-adaptative step}

\item{nbAdaptMarche}{the number of adaptation steps, one step equals 100 iterations}

\item{y}{a vector of count values, from 0 to maxBinom[i] for y[i], i in 1...length(y)}

\item{X}{a matrix of covariates, with dim(X)[1] = length(y)}

\item{maxBinom}{a vector of maximum trials (in binomial fashion), length(maxBinom) = length(y)}

\item{curr}{a vector of initialization values corresponding to each of the covariates in the columns of X, ordered in the same way}

\item{SigMarch}{a covariance matrix for the Gaussian proposal distribution}

\item{prior}{a list of vectors for the priors' hyperparameters}

\item{warmup}{the number of iterations reserved to warmup (warmup >= 0 and warmup <= N)}

\item{invTemp}{a value to temper the likelihood}
}
\value{
a list that includes a matrix of each coefficient posterior sample and the adapted covariance matrix for the Gaussian proposal distribution
}
\description{
Performs logistic binomial regression using the RWMH method to infer the coefficients conditionally
}
