% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{logisticRegression}
\alias{logisticRegression}
\title{Logistic binomial regression / joint inference with RWMH method}
\usage{
logisticRegression(
  form,
  data,
  maxBinom = NULL,
  method = "freq",
  curr = NULL,
  prior = NULL,
  N = 5000,
  nbAdaptMarche = 100,
  warmup = 0,
  epsilon = 1e-07,
  SigMarch = 1,
  invTemp = 1
)
}
\arguments{
\item{form}{an object of class "formula" (as in "lm" or "glm")}

\item{data}{a data.frame containing response variables and covariates}

\item{maxBinom}{a vector of maximum trials (in binomial fashion), length(maxBinom) = nrow(data)}

\item{method}{chosen method for the inference : "freq", "RWMH".}

\item{curr}{a vector of initialization values corresponding to each of the covariates in the chosen columns of data, ordered in the same way}

\item{prior}{a list of vectors for the priors' hyperparameters}

\item{N}{the number of iterations post-adaptative step}

\item{nbAdaptMarche}{the number of adaptation steps, one step equals 100 iterations}

\item{warmup}{the number of iterations reserved to warmup (warmup >= 0 and warmup <= N)}

\item{epsilon}{the number of iterations reserved to warmup (warmup >= 0 and warmup <= N)}

\item{SigMarch}{a covariance matrix for the Gaussian proposal distribution (potential only a scalar)}

\item{invTemp}{a value to temper the likelihood}
}
\value{
a list that includes a matrix of each coefficient posterior sample, the adapted covariance matrix for the Gaussian proposal distribution and some values to compute WAIC
}
\description{
This function performs a logistic binomial regression using the RWMH method to jointly infer the coefficients.
}
\details{
...
}
\examples{
...

}
